backend/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pillow==11.1.0
pillow_heif==0.21.0
PyPDF2==3.0.1
Werkzeug==3.1.3
ffmpeg-python==0.2.0